requirements.txt
huggingface_hub
pypdf
sentence-transformers
llama_index
langchain
torch
chromadb
python-dotenv
-e.